+++ krita/plugins/colorspaces/ctlcs/KoCtlConvolutionOp.h
@@ -31,7 +31,7 @@ class KoCtlConvolutionOp : public KoConvolutionOp
 public:
     KoCtlConvolutionOp(KoCtlColorSpace* _colorSpace, const KoCtlColorSpaceInfo* _info);
     virtual ~KoCtlConvolutionOp();
-    virtual void convolveColors(const quint8* const* colors, const qint32* kernelValues, quint8 *dst, qint32 factor, qint32 offset, qint32 nPixels, const QBitArray & channelFlags) const;
+    virtual void convolveColors(const quint8* const* colors, const qreal* kernelValues, quint8 *dst, qreal factor, qreal offset, qint32 nPixels, const QBitArray & channelFlags) const;
 private:
     QList<KoCtlAccumulator*> m_accumulators;
     KoCtlColorSpace* m_colorSpace;
